requirements.txt
Pillow
cryptography
Flask
rfc3161ng
vobject
iOSbackup
reportlab~=4.3.1
protobuf-decoder
bleach
sqlalchemy~=2.0.40
sqlalchemy_utils
dotenv~=0.9.9
azure.cognitiveservices.speech
azure.ai.vision.imageanalysis
azure.core
pandas~=2.2.3
transformers~=4.51.3
psycopg2
torch
presidio_analyzer~=2.2.358
azure.ai.textanalytics
openai~=1.78.0
SQLAlchemy-Utils~=0.41.2
azure-core~=1.34.0
python-dotenv~=1.1.0
httpx~=0.28.1
wheel